Added :escape_only to the benchmark & tests.
Also changes the Loofah lines to what they consider to be
'semantically clearer' in the distance hope that it would fix
the Loofah benchmark. With my version of the gem (loofah 0.4.7),
it chokes to death on :strip for Slashdot. :prune works fine,
though!

Here's how we stand (with Loofah commented out because
I'm not fixing whatever is wrong with it):

-- Benchmark --
  Slashdot HTML doc (98255 bytes) x 100
                                   total    single    rel
          Sanitize.clean (strip)  34.173 (0.341732)     -
          Sanitize.clean (prune)  14.549 (0.145494)  0.43x
         Sanitize.clean (escape)  56.120 (0.561203)  1.64x

  Big HTML doc (149413 bytes) x 100
                                   total    single    rel
          Sanitize.clean (strip)  19.025 (0.190254)     -
          Sanitize.clean (prune)  11.194 (0.111939)  0.59x
         Sanitize.clean (escape)  33.436 (0.334364)  1.76x

  Small HTML fragment (1510 bytes) x 1000
                                   total    single    rel
          Sanitize.clean (strip)   3.646 (0.003646)     -
          Sanitize.clean (prune)   3.326 (0.003326)  0.91x
         Sanitize.clean (escape)   3.306 (0.003306)  0.91x

  Tiny HTML fragment (39 bytes) x 10000
                                   total    single    rel
          Sanitize.clean (strip)   6.325 (0.000632)     -
          Sanitize.clean (prune)   6.206 (0.000621)  0.98x
         Sanitize.clean (escape)   6.369 (0.000637)  1.01x

---

I didn't wanna fiddle with the main test file
'cuz then I wouldn't be sure if I had broken something
important after.